Fixed postgres floating point type (#952)

+++ pgml-sdks/rust/pgml/src/filter_builder.rs
@@ -117,9 +117,9 @@ fn get_value_type(value: &serde_json::Value) -> String {
     } else if value.is_string() {
         "text".to_string()
     } else if value.is_i64() {
-        "bigint".to_string()
+        "float8".to_string()
     } else if value.is_f64() {
-        "double".to_string()
+        "float8".to_string()
     } else if value.is_boolean() {
         "bool".to_string()
     } else {
